Move random number generation to generators module
