+++ src/api/__init__.py
@@ -1,3 +1,3 @@
 """Public API for our projects"""
 
-__version__ = "1.7.0"
+__version__ = "1.8.0"
