+++ src/api/server.py
@@ -1,12 +1,10 @@
 """API server definition"""
 
 from os import getenv
-from random import randint
 
 import sentry_sdk
 from fastapi import APIRouter, FastAPI
 from imsosorry import uwuify
-
 # pylint: disable-next=no-name-in-module
 from pydantic import BaseModel
 
@@ -31,13 +29,6 @@ class TextModel(BaseModel):
     """Generic model for accepting arbitrary plain-text input"""
 
     text: str
-
-
-class Numbers(BaseModel):
-    """Model to hold a list of numbers"""
-
-    numbers: list[int]
-    total: int
 
 
 app = FastAPI(
@@ -69,13 +60,5 @@ async def uwuify_route(text: TextModel):
     return {"text": uwuify(text.text)}
 
 
-@router_fun.get("/random-numbers/{quantity}/{range_high}")
-async def random_numbers(quantity: int, range_high: int) -> Numbers:
-    """Generate bulk random numbers"""
-    numbers = [randint(1, range_high) for _ in range(quantity)]
-    return Numbers(numbers=numbers, total=sum(numbers))
-
-
 app.include_router(router_fun)
-
 app.include_router(router_generators)
