+++ tests/test_server.py
@@ -15,12 +15,3 @@ def test_read_main():
         "message": "Welcome to the API",
         "version": __version__,
     }
-
-
-def test_read_random_numbers():
-    response = client.get("/fun/random-numbers/1/1")
-    assert response.status_code == 200
-    assert response.json() == {
-        "numbers": [1],
-        "total": 1,
-    }
